refactor docker container related actions.
