+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-containerregistry/src/main/java/com/microsoft/azure/toolkit/intellij/containerregistry/dockerhost/RunOnDockerHostAction.java
@@ -3,12 +3,13 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  */
 
-package com.microsoft.azure.toolkit.intellij.legacy.webapp.action;
+package com.microsoft.azure.toolkit.intellij.containerregistry.dockerhost;
 
 import com.intellij.execution.ProgramRunnerUtil;
 import com.intellij.execution.RunManagerEx;
 import com.intellij.execution.RunnerAndConfigurationSettings;
 import com.intellij.execution.configurations.ConfigurationFactory;
+import com.intellij.execution.configurations.RunConfiguration;
 import com.intellij.execution.executors.DefaultRunExecutor;
 import com.intellij.execution.impl.RunDialog;
 import com.intellij.openapi.actionSystem.AnAction;
@@ -16,10 +17,8 @@ import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.project.Project;
 import com.microsoft.azure.toolkit.intellij.common.IntelliJAzureIcons;
 import com.microsoft.azure.toolkit.intellij.container.model.DockerImage;
-import com.microsoft.azure.toolkit.intellij.legacy.webapp.runner.WebAppConfigurationType;
-import com.microsoft.azure.toolkit.intellij.legacy.webapp.runner.webapponlinux.WebAppOnLinuxDeployConfiguration;
-import com.microsoft.azure.toolkit.lib.common.action.Action;
-import com.microsoft.azure.toolkit.lib.common.action.AzureActionManager;
+import com.microsoft.azure.toolkit.intellij.containerregistry.AzureDockerSupportConfigurationType;
+import com.microsoft.azure.toolkit.intellij.containerregistry.dockerhost.DockerHostRunConfiguration;
 import com.microsoft.azure.toolkit.lib.common.operation.AzureOperation;
 import com.microsoft.azure.toolkit.lib.common.task.AzureTaskManager;
 import com.microsoft.azure.toolkit.lib.containerregistry.Tag;
@@ -28,32 +27,27 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.Optional;
 
-public class WebAppOnLinuxAction extends AnAction {
-    private static final String DIALOG_TITLE = "Deploy Image to Web App";
-    private static final WebAppConfigurationType configType = WebAppConfigurationType.getInstance();
+public class RunOnDockerHostAction extends AnAction {
+    private static final String DIALOG_TITLE = "Run on Docker Host";
+    private static final AzureDockerSupportConfigurationType configType = AzureDockerSupportConfigurationType.getInstance();
     private final DockerImage dockerImage;
 
-    public WebAppOnLinuxAction() {
+    public RunOnDockerHostAction() {
         this(null);
     }
 
-    public WebAppOnLinuxAction(@Nullable DockerImage dockerImage) {
-        //noinspection DialogTitleCapitalization
-        super(DIALOG_TITLE, "Build/push local image to Azure Web App", IntelliJAzureIcons.getIcon("/icons/DockerSupport/RunOnWebApp.svg"));
+    public RunOnDockerHostAction(@Nullable final DockerImage dockerImage) {
+        super(DIALOG_TITLE, "Build image and run in local docker host", IntelliJAzureIcons.getIcon("/icons/DockerSupport/Run.svg"));
         this.dockerImage = dockerImage;
     }
 
     @Override
-    @AzureOperation(name = "user/docker.start_app")
+    @AzureOperation(name = "user/springcloud.deploy_app")
     public void actionPerformed(@Nonnull AnActionEvent e) {
-        final Project project = e.getProject();
-        if (project != null) {
-            AzureActionManager.getInstance().getAction(Action.REQUIRE_AUTH)
-                .handle(() -> runConfiguration(project, this.dockerImage));
-        }
+        Optional.ofNullable(e.getProject()).ifPresent(p -> runConfiguration(p, this.dockerImage));
     }
 
-    public static void deploy(@Nonnull Tag tag, @Nonnull Project project) {
+    public static void run(@Nonnull Tag tag, @Nonnull Project project) {
         final DockerImage image = DockerImage.builder()
             .isDraft(false)
             .repositoryName(tag.getParent().getParent().getName())
@@ -62,15 +56,16 @@ public class WebAppOnLinuxAction extends AnAction {
         runConfiguration(project, image);
     }
 
-    private static void runConfiguration(@Nonnull Project project, @Nullable DockerImage image) {
+    private static void runConfiguration(Project project, DockerImage dockerImage) {
         final RunManagerEx manager = RunManagerEx.getInstanceEx(project);
-        final ConfigurationFactory factory = configType.getWebAppOnLinuxConfigurationFactory();
+        final ConfigurationFactory factory = configType.getDockerHostRunConfigurationFactory();
         final String configurationName = String.format("%s: %s", factory.getName(), project.getName());
         final RunnerAndConfigurationSettings existingSettings = manager.findConfigurationByName(configurationName);
         final RunnerAndConfigurationSettings settings = Optional.ofNullable(existingSettings)
             .orElseGet(() -> manager.createConfiguration(configurationName, factory));
-        if (settings.getConfiguration() instanceof WebAppOnLinuxDeployConfiguration) {
-            ((WebAppOnLinuxDeployConfiguration) settings.getConfiguration()).setDockerImage(image);
+        final RunConfiguration configuration = settings.getConfiguration();
+        if (configuration instanceof DockerHostRunConfiguration) {
+            Optional.ofNullable(dockerImage).ifPresent(image -> ((DockerHostRunConfiguration) configuration).setDockerImage(image));
         }
         AzureTaskManager.getInstance().runLater(() -> {
             if (RunDialog.editConfiguration(project, settings, DIALOG_TITLE, DefaultRunExecutor.getRunExecutorInstance())) {
