+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-containerregistry/src/main/resources/META-INF/azure-intellij-plugin-containerregistry.xml
@@ -22,11 +22,11 @@
                 text="Add Docker Support" description="Add docker support"
                 icon="/icons/DockerSupport/DockerSupport.svg">
         </action>
-        <action id="Actions.RunOnDockerHost" class="com.microsoft.azure.toolkit.intellij.containerregistry.action.RunOnDockerHostAction"
+        <action id="Actions.RunOnDockerHost" class="com.microsoft.azure.toolkit.intellij.containerregistry.dockerhost.RunOnDockerHostAction"
                 text="Docker Run" description="Docker Run"
                 icon="/icons/DockerSupport/Run.svg">
         </action>
-        <action id="Actions.PushImage" class="com.microsoft.azure.toolkit.intellij.containerregistry.action.PushImageAction"
+        <action id="Actions.PushImage" class="com.microsoft.azure.toolkit.intellij.containerregistry.pushimage.PushImageAction"
                 text="Push Image" description="Push Image"
                 icon="/icons/DockerSupport/PushImage.svg">
         </action>
