+++ Utils/azure-toolkit-ide-libs/azure-toolkit-ide-common-lib/src/main/resources/bundles/com/microsoft/azure/toolkit/operation.properties
@@ -342,8 +342,9 @@ platform/sdk.warn_deprecated_libs=warn on deprecated libs
 user/sdk.open_reference_book=open Azure SDK Reference Book
 user/sdk.refresh_dependency=load dependencies of selected project module
 user/sdk.add_dependency=add dependency to selected project module
-user/docker.push_image=push docker image to web app within run/debug configuration
-user/docker.start_app=start docker based web app within run/debug configuration
+user/docker.push_image=push docker image to Azure Container Registry
+user/docker.run_image=run docker image as container
+user/docker.start_app=start docker container based web app
 platform/docker.deploy_image=deploy docker image to web app
 user/function.run_app.module=run function app module ({0}) locally
 platform/connector.update_java_parameters_for_configuration.config=setup resource connection(s) for run configuration ({0})
